Initial work on adding a requester of USB permissions
I was able to include the permission XML that allows any app to "respond" to a just-plugged printer, but......... not more than that. Some changes on the iOS side was done too, but that was easier since it's simply a no-op (and is untested too).
Any Java code I included (even blatantly broken stuff) never caused an issue during compilation or runtime, not more than the current bug causes. No clue how to actually compile the plugin and get it into the debug app.

This is supposed to be a solution for https://github.com/auctifera-josed/starprnt/issues/58